addStrelkaGt skips entries that cannot parse a GT value
Improperly parsed GT values `/`, `/1`, `1/` for example, were causing
data issues downstream. Specifically when merging with other VCFs, the
when a malformed value is encountered, it just stops and moves on to
the next file. This meant many values were silently dropped.

We avoid the issue by simply not including those entries in the output
VCF. Instead of writing the line to file we print a message to stderr
and move on to the next line.