Add custom annotations to VEP, parsed in their own task
VEP has had a TODO on adding custom annotations for a while because of
difficulties working it into WDL. The issue is operating on an Array
of structs. WDL can either scatter over an array in a workflow, or
join array contents into a string in a command IFF those contents are
primitives, not a struct. The resolution to this is the incredibly
dissatisfying "do it in a VM" answer. Scatter over each annotation so
each VM only has to deal with a single struct, and do interpolation on
primitives from there.

In order to hide this implementation from callers, `vep` now refers to
a new workflow, not the original task. The new workflow does the
original task AND the parsing for Array struct. To callers there
should be no difference.