Force bai timestamp to avoid reindex
Cromwell consistently downloads the .bai files before their .bam,
which causes cnvkit to reindex them. This would be not that big of a
deal, except in cnvkit:0.9.7+ that reindex causes an
exception. `touch` resolves this by forcing their timestamp and
avoiding the reindex.

We have to change to a forked cnvkit image which strips out a block
preventing root access because Cromwell places files at
/cromwell_root/ which requires root access to modify, so calls to
touch were failing with permissions errors