Set localization_optional on inputs for some GATK and samtools calls.
Others in the pipeline are still using older versions, so they'll need
to be upgraded first to support this option.

(A couple different implementations are used for grabbing the
authorization toekn based on what tools are available in the existing
docker images.  Notably the busybox version of wget here is missing
several flags that would make the implementations more similar.)